Connection provider - added support for appsettings.json configuration
