+++ SqlRepo.SqlServer.IntegrationTests/SqlRepo.SqlServer.IntegrationTests.csproj
@@ -16,27 +16,29 @@
     <StartupObject />
   </PropertyGroup>
 
+
   <ItemGroup>
     <ProjectReference Include="..\SqlRepo.SqlServer\SqlRepo.SqlServer.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk"
-                      Version="15.5.0" />
-    <PackageReference Include="FluentAssertions"
-                      Version="4.19.4" />
-    <PackageReference Include="NSubstitute"
-                      Version="3.1.0" />
-    <PackageReference Include="NUnit"
-                      Version="3.9.0" />
-    <PackageReference Include="dotnet-test-nunit"
-                      Version="3.4.0-beta-3" />
-    <PackageReference Include="NUnit3TestAdapter"
-                      Version="3.9.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.0.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0" />
+    <PackageReference Include="FluentAssertions" Version="4.19.4" />
+    <PackageReference Include="NSubstitute" Version="3.1.0" />
+    <PackageReference Include="NUnit" Version="3.9.0" />
+    <PackageReference Include="dotnet-test-nunit" Version="3.4.0-beta-3" />
+    <PackageReference Include="NUnit3TestAdapter" Version="3.9.0" />
   </ItemGroup>
 
   <ItemGroup>
     <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="appsettings.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
