+++ SqlRepo.SqlServer.IntegrationTests/appsettings.json
@@ -1,0 +1,5 @@
+﻿{
+    "ConnectionStrings": {
+        "TestingDbConn": "Data Source=(LocalDB)\\MSSQLLocalDB;AttachDbFilename=C:\\_src\\SqlRepo\\SqlRepo.SqlServer.IntegrationTests\\Testing.mdf;Integrated Security=True;Connect Timeout=30"
+    }
+}
