+++ SqlRepo.Tests/SqlRepo.Tests.v3.ncrunchproject
@@ -1,6 +1,8 @@
 ﻿<ProjectConfiguration>
   <Settings>
-    <HiddenComponentWarnings />
+    <HiddenComponentWarnings>
+      <Value>NUnitTestsWithSameName</Value>
+    </HiddenComponentWarnings>
     <PreviouslyBuiltSuccessfully>True</PreviouslyBuiltSuccessfully>
   </Settings>
 </ProjectConfiguration>
